reorganize directories + imports
