+++ src/Components/BundleView/NameColumn.styles.ts
@@ -1,0 +1,6 @@
+import { makeStyles } from '@fluentui/react-components';
+
+export const useStyles = makeStyles({
+  nameCell: { display: 'flex', flexDirection: 'row', alignItems: 'center' },
+  nameLabel: { marginLeft: '10px' }
+});
