+++ src/Components/BundleView/index.ts
@@ -1,0 +1,2 @@
+export * from './BundleView';
+export * from './ColumnInfo';
